pyproject.toml
[tool.poetry]
name = "anno-backend"
version = "0.1"
description = "Backend for Medical Image Annotation Platform"
authors = ["TheOneTwelfth"]


[tool.poetry.dependencies]
python = "~3.8"

attrs = "^19.3.0"
flask = "^1.1.2"
flask-apispec = "^0.8.8"
flask-bcrypt = "^0.7.1"
flask-jwt-extended = "^3.24.1"
flask-migrate = "^2.5.3"
flask-sqlalchemy = "^2.4.1"
marshmallow = "^3.6.0"
pyparsing = "^2.4.7"
psycopg2-binary = "^2.8.5"
requests = "^2.23.0"


[tool.poetry.dev-dependencies]
pytest = "^5.4.2"

pytest-black = "^0.3.9"
pytest-cov = "^2.8.1"
pytest-flake8 = "^1.0.6"
pytest-isort = "^1.0.0"
pytest-mypy = "^0.6.2"


[tool.black]
target-version = ['py38']
skip-string-normalization = true
